+++ float/src/fbig.rs
@@ -44,7 +44,7 @@ use dashu_int::{DoubleWord, IBig};
 /// Other formatting traits will be supported in future.
 ///
 /// ```
-/// # use dashu_int::error::ParseError;
+/// # use dashu_base::ParseError;
 /// # use dashu_float::DBig;
 /// // parsing
 /// let a = DBig::from_parts(123456789.into(), -5);
@@ -161,7 +161,7 @@ impl<R: Round, const B: Word> FBig<R, B> {
     /// # Examples
     ///
     /// ```
-    /// # use dashu_int::error::ParseError;
+    /// # use dashu_base::ParseError;
     /// # use dashu_float::DBig;
     /// use dashu_float::Repr;
     ///
@@ -184,7 +184,7 @@ impl<R: Round, const B: Word> FBig<R, B> {
     /// # Examples
     ///
     /// ```
-    /// # use dashu_int::error::ParseError;
+    /// # use dashu_base::ParseError;
     /// # use dashu_float::DBig;
     /// use dashu_base::Approximation::*;
     ///
@@ -232,7 +232,7 @@ impl<R: Round, const B: Word> FBig<R, B> {
     /// # Examples
     ///
     /// ```
-    /// # use dashu_int::error::ParseError;
+    /// # use dashu_base::ParseError;
     /// # use dashu_float::DBig;
     /// let a = DBig::from_parts((-1234).into(), -2);
     /// assert_eq!(a, DBig::from_str_native("-12.34")?);
@@ -258,7 +258,7 @@ impl<R: Round, const B: Word> FBig<R, B> {
     /// # Examples
     ///
     /// ```
-    /// # use dashu_int::error::ParseError;
+    /// # use dashu_base::ParseError;
     /// # use dashu_float::DBig;
     /// use dashu_base::Sign;
     ///
@@ -329,7 +329,7 @@ impl<R: Round, const B: Word> FBig<R, B> {
     /// # Examples
     ///
     /// ```
-    /// # use dashu_int::error::ParseError;
+    /// # use dashu_base::ParseError;
     /// # use dashu_float::DBig;
     /// assert_eq!(DBig::from_str_native("1.23")?.ulp(), DBig::from_str_native("0.01")?);
     /// assert_eq!(DBig::from_str_native("01.23")?.ulp(), DBig::from_str_native("0.001")?);