+++ float/tests/io.rs
@@ -1,7 +1,7 @@
 use core::str::FromStr;
-use dashu_base::Sign;
+use dashu_base::{Sign, ParseError};
 use dashu_float::{round::mode, DBig, FBig};
-use dashu_int::{error::ParseError, DoubleWord, IBig, Word};
+use dashu_int::{DoubleWord, IBig, Word};
 
 mod helper_macros;
 type FBin = FBig;