+++ integer/src/ibig.rs
@@ -27,7 +27,8 @@ use crate::{
 ///
 /// ```
 /// // parsing
-/// # use dashu_int::{error::ParseError, IBig, Word};
+/// # use dashu_base::ParseError;
+/// # use dashu_int::{IBig, Word};
 /// let a = IBig::from(408580953453092208335085386466371u128);
 /// let b = IBig::from(-0x1231abcd4134i64);
 /// let c = IBig::from_str_radix("a2a123bbb127779cccc123", 32)?;