+++ rational/src/error.rs
@@ -1,3 +1,5 @@
+//! Definitions of panic cases
+
 pub const fn panic_divide_by_0() -> ! {
     panic!("Divisor or denominator must not be zero!")
 }